# Conflicts:
#	ARKBreedingStats/_manifest.json
#	ARKBreedingStats/_manifest.tt
#	ARKBreedingStats/json/values/values.json
#	ARKBreedingStats/local/strings.zh.resx
#	ARKBreedingStats/settings/Settings.Designer.cs
#	speciesImages/speciesImages.zip

+++ ARKBreedingStats/FileService.cs
@@ -1,5 +1,6 @@
 ﻿using System;
 using System.IO;
+using System.Net.Http;
 using System.Reflection;
 using System.Runtime.Serialization;
 
@@ -117,7 +118,7 @@ namespace ARKBreedingStats
                     if (data != null)
                         return true;
 
-                    errorMessage = $"File\n{Path.GetFullPath(filePath)}\n contains no readable data.";
+                    errorMessage = $"File\n{Path.GetFullPath(filePath)}\ncontains no readable data.";
                     return false;
                 }
             }
@@ -158,7 +159,7 @@ namespace ARKBreedingStats
         /// <returns>True if the file is not existing after this method ends.</returns>
         public static bool TryDeleteFile(string filePath)
         {
-            if (!File.Exists(filePath)) return true;
+            if (string.IsNullOrEmpty(filePath) || !File.Exists(filePath)) return true;
             try
             {
                 File.Delete(filePath);
@@ -279,6 +280,21 @@ namespace ARKBreedingStats
             //    return true;
             //}
             //catch { return false; }
+        }
+
+        private static HttpClient _httpClient;
+
+        /// <summary>
+        /// Returns a static HttpClient. It's apparently better to reuse one object per app only.
+        /// </summary>
+        public static HttpClient GetHttpClient
+        {
+            get
+            {
+                if (_httpClient == null)
+                    _httpClient = new HttpClient();
+                return _httpClient;
+            }
         }
     }
 }
